Added Player Movement
Added player movement mechanics with player controller VSC file, as well as a rigidbody component to the player and some events in another component that I had to plug the player into to get the input controls from.
(Sorry if that doesn't make sense - watch episode 1 of Game Code Library's top down unity 2D series to make it make sense).